+++ backend/dataall/api/Objects/DataPipeline/resolvers.py
@@ -24,23 +24,33 @@ def create_pipeline(context: Context, source, input=None):
             data=input,
             check_perm=True,
         )
+        if input['devStrategy'] == 'cdk-trunk':
+            Stack.create_stack(
+                session=session,
+                environment_uri=pipeline.environmentUri,
+                target_type='cdkrepo',
+                target_uri=pipeline.DataPipelineUri,
+                target_label=pipeline.label,
+                payload={'account': pipeline.AwsAccountId, 'region': pipeline.region},
+            )
 
-        Stack.create_stack(
-            session=session,
-            environment_uri=pipeline.environmentUri,
-            target_type='pipeline',
-            target_uri=pipeline.DataPipelineUri,
-            target_label=pipeline.label,
-            payload={'account': pipeline.AwsAccountId, 'region': pipeline.region},
-        )
-        Stack.create_stack(
-            session=session,
-            environment_uri=pipeline.environmentUri,
-            target_type='pipelinePip',
-            target_uri=f"{pipeline.DataPipelineUri}pip",
-            target_label=pipeline.label,
-            payload={'account': pipeline.AwsAccountId, 'region': pipeline.region},
-        )
+            Stack.create_stack(
+                session=session,
+                environment_uri=pipeline.environmentUri,
+                target_type='cdkpipeline',
+                target_uri=f"{pipeline.DataPipelineUri}pip",
+                target_label=pipeline.label,
+                payload={'account': pipeline.AwsAccountId, 'region': pipeline.region},
+            )
+        else:
+            Stack.create_stack(
+                session=session,
+                environment_uri=pipeline.environmentUri,
+                target_type='pipeline',
+                target_uri=pipeline.DataPipelineUri,
+                target_label=pipeline.label,
+                payload={'account': pipeline.AwsAccountId, 'region': pipeline.region},
+            )
 
     stack_helper.deploy_stack(context, pipeline.DataPipelineUri)
