+++ backend/dataall/cdkproxy/cdk_cli_wrapper.py
@@ -54,7 +54,7 @@ def update_stack_output(session, stack):
         stack.outputs = outputs
 
 
-def clone_remote_stack(pipeline, pipeline_environment): #, development_environments):
+def clone_remote_stack(pipeline, pipeline_environment):
     print('..................................................')
     print('     Configure remote CDK app                     ')
     print('..................................................')
@@ -212,15 +212,14 @@ def deploy_cdk_stack(engine: Engine, stackid: str, app_path: str = None):
                 stack.stackid = meta['StackId']
                 stack.status = meta['StackStatus']
                 update_stack_output(session, stack)
-                if stack.stack == 'pipeline':
+                if stack.stack == 'cdkrepo':
                     logger.warning(f'Starting new remote stack from  targetUri {stack.targetUri}pip')
                     cicdstack: models.Stack = Stack.get_stack_by_target_uri(session, target_uri=f"{stack.targetUri}pip")
                     cicdstack.EcsTaskArn = stack.EcsTaskArn
                     session.commit()
                     pipeline = Pipeline.get_pipeline_by_uri(session, stack.targetUri)
                     pipeline_environment = Environment.get_environment_by_uri(session, pipeline.environmentUri)
-                    #development_environments = Pipeline.query_pipeline_environments(session, pipeline.DataPipelineUri)
-                    clone_remote_stack(pipeline, pipeline_environment) #, development_environments)
+                    clone_remote_stack(pipeline, pipeline_environment)
                     deploy_cdk_stack(engine, cicdstack.stackUri, f"./stacks/{pipeline.repo}/app.py")
             else:
                 stack.status = 'CREATE_FAILED'
