+++ frontend/src/views/Pipelines/PipelineListItem.js
@@ -199,7 +199,7 @@ const PipelineListItem = (props) => {
             </Grid>
             <Grid item md={8} xs={12}>
               <Typography color="textPrimary" variant="body2">
-                <StackStatus status={pipeline.cicdStack?.status} />
+                <StackStatus status={pipeline.devStrategy == "cdk-trunk" ? pipeline.cicdStack.status : pipeline.stack.status} />
               </Typography>
             </Grid>
           </Grid>
